rename to errno_handling.h
