+++ src/support/errno_handling.h
