+++ src/support/socket.h
@@ -55,7 +55,7 @@
 #include <unordered_map>
 #include <vector>
 
-#include "error_handling.h"
+#include "errno_handling.h"
 #include "ssize.h"
 #include "utils.h"
